+++ src/Client.php
@@ -9,6 +9,8 @@ namespace Jahuty;
  */
 class Client
 {
+    private $cache;
+
     private $client;
 
     private $key;
@@ -17,11 +19,19 @@ class Client
 
     private $resources;
 
+    private $options = [
+        'cache'    => null,
+        'ttl'      => null,
+        'base_uri' => Jahuty::BASE_URI
+    ];
+
     private $services;
 
-    public function __construct(string $key)
+    public function __construct(string $key, array $options = [])
     {
         $this->key = $key;
+
+        $this->setOptions($options);
     }
 
     public function __get(string $name): Service\Service
@@ -33,10 +43,23 @@ class Client
         return $this->services->$name;
     }
 
+    public function fetch(Action\Action $action, Ttl\Ttl $ttl): Resource\Resource
+    {
+        if (null === $this->cache) {
+            $this->cache = new Cache\Manager(
+                $this,
+                $this->getOption('cache') ?: new Cache\Memory(),
+                $this->getOption('ttl')
+            );
+        }
+
+        return $this->cache->fetch($action, $ttl);
+    }
+
     public function request(Action\Action $action): Resource\Resource
     {
         if (null === $this->requests) {
-            $this->requests = new Request\Factory();
+            $this->requests = new Request\Factory($this->getOption('base_uri'));
         }
 
         $request = $this->requests->new($action);
@@ -58,5 +81,34 @@ class Client
         }
 
         return $resource;
+    }
+
+    private function getOption(string $name)
+    {
+        if (!\array_key_exists($name, $this->options)) {
+            throw new \OutOfBoundsException(
+                "Option '$name' does not exist"
+            );
+        }
+
+        return $this->options[$name];
+    }
+
+    private function setOptions(array $options): void
+    {
+        $options = \array_merge($this->options, $options);
+
+        if ($options['cache'] !== null &&
+            !($options['cache'] instanceof \Psr\SimpleCache\CacheInterface)
+        ) {
+            throw new \InvalidArgumentException(
+                "Option 'cache' must be null or CacheInterface"
+            );
+        }
+
+        // Accepts null, int, or DateInterval.
+        $options['ttl'] = new Ttl\Ttl($options['ttl']);
+
+        $this->options = $options;
     }
 }