+++ src/Service/Snippet.php
@@ -4,19 +4,30 @@ namespace Jahuty\Service;
 
 use Jahuty\Action\Show;
 use Jahuty\Resource\Resource;
+use Jahuty\Ttl\Ttl;
 
 class Snippet extends Service
 {
     public function render(int $id, array $options = []): Resource
     {
-        $params = [];
+        $defaults = [
+            'params' => null,
+            'ttl'    => null
+        ];
 
-        if (\array_key_exists('params', $options)) {
-            $params['params'] = \json_encode($options['params'], JSON_THROW_ON_ERROR);
+        $options = \array_merge($defaults, $options);
+
+        $params = [];
+        if ($options['params']) {
+            $params['params'] = \json_encode(
+                $options['params'],
+                JSON_THROW_ON_ERROR
+            );
         }
 
         $action = new Show('render', $id, $params);
+        $ttl    = new Ttl($options['ttl']);
 
-        return $this->client->request($action);
+        return $this->client->fetch($action, $ttl);
     }
 }