+++ tests/Service/SnippetTest.php
@@ -4,6 +4,7 @@ namespace Jahuty\Service;
 
 use Jahuty\Client;
 use Jahuty\Action\Show;
+use Jahuty\Ttl\Ttl;
 
 class SnippetTest extends \PHPUnit\Framework\TestCase
 {
@@ -13,8 +14,8 @@ class SnippetTest extends \PHPUnit\Framework\TestCase
 
         $client = $this->createMock(Client::class);
         $client->expects($this->once())
-            ->method('request')
-            ->with($this->equalTo($action));
+            ->method('fetch')
+            ->with($this->equalTo($action), $this->isInstanceOf(Ttl::class));
 
         (new Snippet($client))->render(1);
     }
@@ -25,9 +26,24 @@ class SnippetTest extends \PHPUnit\Framework\TestCase
 
         $client = $this->createMock(Client::class);
         $client->expects($this->once())
-            ->method('request')
-            ->with($this->equalTo($action));
+            ->method('fetch')
+            ->with($this->equalTo($action), $this->isInstanceOf(Ttl::class));
 
         (new Snippet($client))->render(1, ['params' => ['foo' => 'bar']]);
+    }
+
+    public function testRenderWhenParamsAndTllDoExist(): void
+    {
+        $action = new Show('render', 1, ['params' => '{"foo":"bar"}']);
+
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('fetch')
+            ->with($this->equalTo($action), $this->equalTo(new Ttl(60)));
+
+        (new Snippet($client))->render(1, [
+            'params' => ['foo' => 'bar'],
+            'ttl'    => 60
+        ]);
     }
 }