+++ CHANGELOG.md
@@ -3,11 +3,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 5.1.0 - 2020-12-26
+
+- Add a second, optional parameter to `Jahuty\Client`, an options array with support for the following options: `cache`, a `CacheInterface` implementation); `ttl`, a `null`, `int`, or `DateInterval` SDK-wide time-to-live; and, `base_uri`, the base URI for API requests (useful for testing).
+- Add `fetch()` to `Jahuty\Client`, which checks the cache before requesting an action.
+- Add `Jahuty\Cache\Memory`, a PSR-16 compliant in-memory cache adapter.
+- Change `Jahuty\Client` to default to the in-memory cache, if no `cache` is configured.
+- Move system tests to a separate test case, and improve the testing of `Jahuty\Client`.
+
 ## 5.0.0 - 2020-12-23
 
 - Reduce root namespace from `Jahuty\Jahuty` to `Jahuty`.
 - Change from a static-based architecture (e.g., `\Jahuty\Jahuty\Snippet::render(1)`) to an instance-based one (e.g., `$jahuty->snippets->render(1)`) to make the library easier to develop, test, and use.
 - Fix links in `README`.
+- Add code coverage analysis.
 
 ## 4.0.0 - 2020-07-04